Move header and local file message components into their own files
